ai_service/requirements.txt
# Temel Bağımlılıklar
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cssselect==1.3.0
exceptiongroup==1.3.0
fastapi==0.115.12
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
greenlet==3.2.2
h11==0.16.0
httptools==0.6.4
httpx==0.27.0
idna==3.10
jieba3k==0.35.1
joblib==1.5.0
lxml==5.4.0
lxml_html_clean==0.4.2
newspaper3k==0.2.8
nltk==3.9.1
pillow==11.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1

# AI/ML ve Veritabanı Bağımlılıkları
spacy==3.7.4
sentence-transformers==2.5.1
torch==2.2.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2

# Veritabanı
psycopg2-binary==2.9.10
pgvector==0.2.5

# Vector İşlemleri
faiss-cpu==1.7.4
numpy==1.26.3

# Graf Analizi ve Topluluk Tespiti
networkx==3.3
python-louvain==0.16

# LLM API ve Hata Toleransı
google-generativeai==0.3.2
tenacity==8.2.3
